Move the classes defining spaces, margins, and paddings in the app to Space.kt
