+++ Jetcaster/tv-app/src/main/java/com/example/jetcaster/tv/ui/component/ErrorState.kt
@@ -32,7 +32,7 @@ import androidx.tv.material3.ExperimentalTvMaterial3Api
 import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Text
 import com.example.jetcaster.tv.R
-import com.example.jetcaster.tv.ui.JetcasterAppDefaults
+import com.example.jetcaster.tv.ui.theme.JetcasterAppDefaults
 
 @OptIn(ExperimentalTvMaterial3Api::class)
 @Composable
